requirements.txt
flask==2.3.3
pydantic==2.9.2
groq==0.11.0
aiofiles==23.2.1
python-dotenv==1.0.1
httpx==0.27.2